+++ action/GlaneurGlanerFichierExemple.class.php
@@ -1,6 +1,6 @@
 <?php
 
-class GlaneurLocalGlaner extends ActionExecutor {
+class GlaneurGlanerFichierExemple extends ActionExecutor {
 
     /**
      * @throws Exception
@@ -8,9 +8,10 @@ class GlaneurLocalGlaner extends ActionExecutor {
     public function go(){
         /** @var GlaneurLocal $connecteur */
         $connecteur = $this->getMyConnecteur();
-        $result = $connecteur->glaner();
+
+        $connecteur->glanerFicExemple();
         $this->setLastMessage(implode("<br/>",$connecteur->getLastMessage()));
-        return $result;
+        return true;
     }
 
 }